requirements.txt
CacheControl==0.12.5
cachetools==3.1.1
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
cryptography==2.8
Django==2.2.7
google-api-core==1.15.0
google-api-python-client==1.7.11
google-auth==1.9.0
google-auth-httplib2==0.0.3
google-cloud-core==1.1.0
google-cloud-firestore==1.6.0
google-cloud-storage==1.23.0
google-resumable-media==0.5.0
googleapis-common-protos==1.6.0
grpcio==1.25.0
httplib2==0.14.0
idna==2.8
jwcrypto==0.6.0
msgpack==0.6.2
Naked==0.1.31
numpy==1.17.4
oauth2client==4.1.3
pandas==0.25.3
protobuf==3.11.1
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycparser==2.19
python-dateutil==2.8.1
python-firebase==1.2
pytz==2019.3
PyYAML==5.2
requests==2.22.0
rsa==4.0
shellescape==3.4.1
six==1.13.0
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.25.7
xlrd==1.2.0
gunicorn==20.0.4